backend/requirements.txt
# Core dependencies
gymnasium==1.2.1
networkx==3.5
numpy==2.3.3
pyyaml==6.0.2
scipy==1.16.2

# OSM and geospatial
osmnx==2.0.6
geopandas==1.1.1
shapely==2.1.2
pyproj==3.7.2
pyogrio==0.11.1
rasterio==1.4.3

# Visualization
matplotlib==3.10.6
seaborn==0.13.2

# Parallel processing
joblib==1.5.2

# Testing
pytest==8.4.2
pytest-cov==4.1.0

pandas==2.3.2
tqdm==4.67.1

fastapi
uvicorn
numpy
pydantic
pyyaml